+++ web-app/src/components/car-description/CarDescription.js
@@ -8,6 +8,7 @@ const css = `
     display: flex;
     flex-direction: row;
     justify-content: space-between;
+    align-items: center;
 }
 `;
