+++ web-app/src/components/managerContactDetails/BlockFordCardsAndManagerContactDetails.js
@@ -11,6 +11,15 @@ const css = `
   }`;
 
 export default function PooledСomponent() {
+  const box = {
+    flexGrow: 1,
+    bgcolor: "background.paper",
+    display: "flex",
+    width: 616,
+    height: 260,
+    marginRight: 5,
+    padding: 0,
+  };
   const picturers = {
     pic1: "/img/main_pic5.jpg",
     pic2: "/img/main_pic501.jpg",
@@ -21,7 +30,7 @@ export default function PooledСomponent() {
     <div className="App">
       <div className="flex">
         <style type="text/css">{css}</style>
-        <BlockFordCards hull_types={picturers} />
+        <BlockFordCards sx={box} hull_types={picturers} />
         <ManagerContactDetails />
       </div>
     </div>
