+++ web-app/src/pages/configurator_p1/ConfiguratorP1.js
@@ -4,7 +4,7 @@ import React from "react";
 import AppHeader from "../../components/app-header/AppHeader";
 import BreadCrumbs from "../../components/breadcrumb/Breadcrumb";
 import Configurator from "../../components/Configurator";
-import BlockFordCardsNext from "../../components/BlockFordCardsNext";
+import BlockFordCards from "../../components/BlockFordCards";
 import PickingSwitch from "../../components/picking_switch/PickingSwitch";
 import store from "../../store/store";
 import CarDescription from "../../components/car-description/CarDescription";
@@ -27,7 +27,7 @@ export default function ConfiguratorP1() {
       <Configurator />
       <div className="App">
         <CarDescription />
-        <BlockFordCardsNext sx={box} hull_types={hull_types.hullTypes} />
+        <BlockFordCards sx={box} hull_types={hull_types.hullTypes} />
       </div>
       <PickingSwitch />
     </div>
